refactor(cli): extract prompt-history recall into its own hook
Review request from @vivere-dally: app.tsx should compose hooks and define
layout, not own a business vertical.

The recall layer moves to hooks/prompt_recall.ts, and moves as a HOOK rather
than a relocated factory — `usePromptRecall(composer)` owns the recall position,
registers the layer, and returns only the `canRecall` affordance state its host
renders. app.tsx now composes it in one line and holds no recall state at all,
matching how watchSidebarData and watchProfileParity are already consumed. The
composer arrives as a getter, since the textarea ref is assigned after the hook
runs and the layer re-reads it every keystroke.

`historyRecallLayer` stays exported from the hook module so the dispatch test
keeps driving the SAME config the hook installs over injected seams, never a
replica; the test moves to hooks/prompt_recall.render.test.tsx beside it.

The set-text + caret-to-end pair the retract seed and recall share is now
`seedTextareaText` in components/text_area.tsx — with callers in two modules it
is genuinely multi-caller, so a shared home is justified where duplicating it
would not be.

No behavior change: the full TUI suite is green at 509, and app.tsx is down from
1215 to 1003 lines.
